requirements.txt
python-dotenv==0.14.0

# Database
braintree==4.4.0
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
psycopg2==2.8.6

# Test and Debug
pytest==6.1.0
ipdb==0.13.3
